Bugfix for bad improper assignment in core.qtopology._get_improper
Residue index variable "i" (epic choice for var.name) was overwritten later in
the loop.
It passed the topology building test because it annoyingly produced the exact
same amount of impropers as Q (102), although the impropers were actually different.
(missing impropers between neighboring residues and extra impropers between
non-neighboring residues)
The bug affects only impropers that include head/tail.

The energies are now the same as in Q.
The tests are updated.