Update q_plot.py to specify location argument
Presumably matplotlib has changed since this code was originally written, and I can't see a version of matplotlib that this is intended for. It all seems to work fine in recent versions, except it throws an error from providing too many position arguments. Specifying the position is for the location fixes this, and allows changing between different graphs without closing and reopening the program.